+++ lib/simple_native_image_compress.dart
@@ -1,2 +1,2 @@
-export 'src/bridge_definitions.dart' show CompressFormat;
+export 'src/bridge_definitions.dart' show CompressFormat, FilterType;
 export 'src/simple_native_image_compress.dart';